js: moved all panes to js/panes/
