+++ js/panes/PropertiesPane.js
